requirements.txt
Flask==2.1.0
psycopg2==2.9.3
pytest==6.2.5
python-dotenv